functions/retrieve_kb/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.14.0
google-genai==1.3.0
